move signup from users to account_controller
There was a lot of special case handling going on in the users_controller
for this. Lot simpler this way.
